// frontend/src/api/generated/models/AuthLocalPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: contact-email@something.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthLocalPostRequest
 */
export interface AuthLocalPostRequest {
    /**
     * 
     * @type {string}
     * @memberof AuthLocalPostRequest
     */
    identifier?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthLocalPostRequest
     */
    password?: string;
}

/**
 * Check if a given object implements the AuthLocalPostRequest interface.
 */
export function instanceOfAuthLocalPostRequest(value: object): value is AuthLocalPostRequest {
    return true;
}

export function AuthLocalPostRequestFromJSON(json: any): AuthLocalPostRequest {
    return AuthLocalPostRequestFromJSONTyped(json, false);
}

export function AuthLocalPostRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthLocalPostRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'identifier': json['identifier'] == null ? undefined : json['identifier'],
        'password': json['password'] == null ? undefined : json['password'],
    };
}

export function AuthLocalPostRequestToJSON(json: any): AuthLocalPostRequest {
    return AuthLocalPostRequestToJSONTyped(json, false);
}

export function AuthLocalPostRequestToJSONTyped(value?: AuthLocalPostRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'identifier': value['identifier'],
        'password': value['password'],
    };
}