// frontend/src/api/generated/models/AuthLocalRegisterPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: contact-email@something.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthLocalRegisterPostRequest
 */
export interface AuthLocalRegisterPostRequest {
    /**
     * 
     * @type {string}
     * @memberof AuthLocalRegisterPostRequest
     */
    username?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthLocalRegisterPostRequest
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthLocalRegisterPostRequest
     */
    password?: string;
}

/**
 * Check if a given object implements the AuthLocalRegisterPostRequest interface.
 */
export function instanceOfAuthLocalRegisterPostRequest(value: object): value is AuthLocalRegisterPostRequest {
    return true;
}

export function AuthLocalRegisterPostRequestFromJSON(json: any): AuthLocalRegisterPostRequest {
    return AuthLocalRegisterPostRequestFromJSONTyped(json, false);
}

export function AuthLocalRegisterPostRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthLocalRegisterPostRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'username': json['username'] == null ? undefined : json['username'],
        'email': json['email'] == null ? undefined : json['email'],
        'password': json['password'] == null ? undefined : json['password'],
    };
}

export function AuthLocalRegisterPostRequestToJSON(json: any): AuthLocalRegisterPostRequest {
    return AuthLocalRegisterPostRequestToJSONTyped(json, false);
}

export function AuthLocalRegisterPostRequestToJSONTyped(value?: AuthLocalRegisterPostRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'username': value['username'],
        'email': value['email'],
        'password': value['password'],
    };
}