// frontend/src/api/generated/models/AuthForgotPasswordPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: contact-email@something.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthForgotPasswordPostRequest
 */
export interface AuthForgotPasswordPostRequest {
    /**
     * 
     * @type {string}
     * @memberof AuthForgotPasswordPostRequest
     */
    email?: string;
}

/**
 * Check if a given object implements the AuthForgotPasswordPostRequest interface.
 */
export function instanceOfAuthForgotPasswordPostRequest(value: object): value is AuthForgotPasswordPostRequest {
    return true;
}

export function AuthForgotPasswordPostRequestFromJSON(json: any): AuthForgotPasswordPostRequest {
    return AuthForgotPasswordPostRequestFromJSONTyped(json, false);
}

export function AuthForgotPasswordPostRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthForgotPasswordPostRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'email': json['email'] == null ? undefined : json['email'],
    };
}

export function AuthForgotPasswordPostRequestToJSON(json: any): AuthForgotPasswordPostRequest {
    return AuthForgotPasswordPostRequestToJSONTyped(json, false);
}

export function AuthForgotPasswordPostRequestToJSONTyped(value?: AuthForgotPasswordPostRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'email': value['email'],
    };
}