// frontend/src/api/generated/models/AuthResetPasswordPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: contact-email@something.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthResetPasswordPostRequest
 */
export interface AuthResetPasswordPostRequest {
    /**
     * 
     * @type {string}
     * @memberof AuthResetPasswordPostRequest
     */
    password?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthResetPasswordPostRequest
     */
    passwordConfirmation?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthResetPasswordPostRequest
     */
    code?: string;
}

/**
 * Check if a given object implements the AuthResetPasswordPostRequest interface.
 */
export function instanceOfAuthResetPasswordPostRequest(value: object): value is AuthResetPasswordPostRequest {
    return true;
}

export function AuthResetPasswordPostRequestFromJSON(json: any): AuthResetPasswordPostRequest {
    return AuthResetPasswordPostRequestFromJSONTyped(json, false);
}

export function AuthResetPasswordPostRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthResetPasswordPostRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'password': json['password'] == null ? undefined : json['password'],
        'passwordConfirmation': json['passwordConfirmation'] == null ? undefined : json['passwordConfirmation'],
        'code': json['code'] == null ? undefined : json['code'],
    };
}

export function AuthResetPasswordPostRequestToJSON(json: any): AuthResetPasswordPostRequest {
    return AuthResetPasswordPostRequestToJSONTyped(json, false);
}

export function AuthResetPasswordPostRequestToJSONTyped(value?: AuthResetPasswordPostRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'password': value['password'],
        'passwordConfirmation': value['passwordConfirmation'],
        'code': value['code'],
    };
}