// frontend/src/api/generated/models/AuthChangePasswordPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: contact-email@something.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthChangePasswordPostRequest
 */
export interface AuthChangePasswordPostRequest {
    /**
     * 
     * @type {string}
     * @memberof AuthChangePasswordPostRequest
     */
    password: string;
    /**
     * 
     * @type {string}
     * @memberof AuthChangePasswordPostRequest
     */
    currentPassword: string;
    /**
     * 
     * @type {string}
     * @memberof AuthChangePasswordPostRequest
     */
    passwordConfirmation: string;
}

/**
 * Check if a given object implements the AuthChangePasswordPostRequest interface.
 */
export function instanceOfAuthChangePasswordPostRequest(value: object): value is AuthChangePasswordPostRequest {
    if (!('password' in value) || value['password'] === undefined) return false;
    if (!('currentPassword' in value) || value['currentPassword'] === undefined) return false;
    if (!('passwordConfirmation' in value) || value['passwordConfirmation'] === undefined) return false;
    return true;
}

export function AuthChangePasswordPostRequestFromJSON(json: any): AuthChangePasswordPostRequest {
    return AuthChangePasswordPostRequestFromJSONTyped(json, false);
}

export function AuthChangePasswordPostRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthChangePasswordPostRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'password': json['password'],
        'currentPassword': json['currentPassword'],
        'passwordConfirmation': json['passwordConfirmation'],
    };
}

export function AuthChangePasswordPostRequestToJSON(json: any): AuthChangePasswordPostRequest {
    return AuthChangePasswordPostRequestToJSONTyped(json, false);
}

export function AuthChangePasswordPostRequestToJSONTyped(value?: AuthChangePasswordPostRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'password': value['password'],
        'currentPassword': value['currentPassword'],
        'passwordConfirmation': value['passwordConfirmation'],
    };
}